% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerGram_gui.R
\name{CancerGram_gui}
\alias{CancerGram_gui}
\title{CancerGram Graphical User Interface}
\usage{
CancerGram_gui()
}
\value{
No return value, called for side effects.
}
\description{
Launches graphical user interface that predicts presence of 
anticancer peptides.
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

\seealso{
\code{\link[shiny]{runApp}}
}
