% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_CancerGramModel}
\alias{install_CancerGramModel}
\title{Install CancerGramModel package containing model for ACP prediction}
\usage{
install_CancerGramModel()
}
\description{
Installs CancerGramModel package containing model required for prediction
of anticancer peptides. Due to large size of our model and file size 
limit on CRAN, it needs to be stored in the external repository.
See readme for more information or in case of installation problems.
}
