% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_txt.R
\name{read_txt}
\alias{read_txt}
\title{Read sequences from .txt file}
\usage{
read_txt(connection)
}
\arguments{
\item{connection}{a \code{\link{connection}} to the text (.txt) file.}
}
\value{
a list of sequences.
}
\description{
Read sequence data in a fasta format saved in text file.
}
\details{
The input file should contain one or more amino acid sequences separated by 
empty line(s) in a fasta format.
}
\examples{
(sequences <- read_txt(system.file("CancerGram/prots.txt", package = "CancerGram")))
}
\keyword{manip}
